renamed exercise 3 v1 (non-codesandbox version) folder
