+++ hw3/Ex-03-create-react-app-vscode/src/components/CharacterCard.js
