+++ hw3/Ex-03-create-react-app-vscode/src/utils/chartColors.js
